+++ requake/plot/map_families.py
@@ -12,12 +12,13 @@ Plot families on a map.
 import logging
 import matplotlib as mpl
 import matplotlib.pyplot as plt
-from matplotlib import cm
-from matplotlib import colors
+import matplotlib.patheffects as patheffects
 import cartopy.crs as ccrs
 import cartopy.feature as cfeature
 from obspy.geodetics import gps2dist_azimuth
-from .plot_utils import plot_title, hover_annotation, duration_string
+from .plot_utils import (
+    plot_title, hover_annotation, duration_string, family_colors, plot_colorbar
+)
 from .cached_tiler import CachedTiler
 from .map_tiles import (
     EsriHillshade,
@@ -115,10 +116,13 @@ def map_families(config):
         rq_exit(1)
     fig, ax = _make_basemap(config)
     trans = ccrs.PlateCarree()
-    cmap = mpl.colormaps['tab10']
-    norm = colors.Normalize(vmin=-0.5, vmax=9.5)
+    try:
+        fcolors, norm, cmap = family_colors(config, families)
+    except ValueError as e:
+        logger.error(e)
+        rq_exit(1)
     trace_ids = []
-    for family in families:
+    for family, color in zip(families, fcolors):
         if family.trace_id not in trace_ids and family.trace_id is not None:
             trace_ids.append(family.trace_id)
         fn = family.number
@@ -128,15 +132,16 @@ def map_families(config):
             f'Family {fn}\n{nevents} evts\n{duration_str}\n'
             f'Z {family.depth:.1f} km'
         )
+        # draw markers using a double border:
+        # white, then black (through path_effects)
+        pe = [patheffects.withStroke(linewidth=2, foreground='k')]
         marker = ax.scatter(
             family.lon, family.lat,
             marker='o', s=100,
-            color=cmap(norm(fn % 10)), edgecolor='k',
+            linewidths=1, facecolor=color, edgecolor='w', path_effects=pe,
             transform=trans, label=label, zorder=10)
         marker.to_annotate = True
-    sm = cm.ScalarMappable(cmap=cmap, norm=norm)
-    cbar = fig.colorbar(sm, ticks=range(10), pad=0.1, ax=ax)
-    cbar.ax.set_ylabel('family number (last digit)')
+    plot_colorbar(config, fig, ax, cmap, norm)
     plot_title(
         ax, len(families), trace_ids, vertical_position=1.05, fontsize=10)