+++ requake/plot/plot_utils.py
@@ -9,10 +9,14 @@ Plot utils.
     GNU General Public License v3.0 or later
     (https://www.gnu.org/licenses/gpl-3.0-standalone.html)
 """
+import contextlib
 import matplotlib.dates as mdates
+import numpy as np
+from matplotlib import cm, colors
+from .colormaps import cmaps
 
 
-def format_time_axis(ax, which='xaxis'):
+def format_time_axis(ax, which='xaxis', grid=True):
     """
     Format the time axis of a Matplotlib plot.
     """
@@ -41,8 +45,9 @@ def format_time_axis(ax, which='xaxis'):
         axis.set_major_locator(_major_locator)
         axis.set_major_formatter(_major_fmt)
         axis.set_minor_locator(_minor_locator)
-        axis.grid(True, which='major', linestyle='--', color='0.5')
-        axis.grid(True, which='minor', linestyle=':', color='0.8')
+        if grid:
+            axis.grid(True, which='major', linestyle='--', color='0.5')
+            axis.grid(True, which='minor', linestyle=':', color='0.8')
     ax.figure.canvas.draw()
 
 
@@ -107,10 +112,15 @@ def hover_annotation(event):
         if cont:
             if hover_on == 'lines':
                 color = element.get_color()
-            else:
+            elif hover_on == 'markers':
                 color = element.get_facecolor()[0]
             element.set_linewidth(3)
             annot.xy = (event.xdata, event.ydata)
+            # set a color contrasting with the element color
+            contrast_color = 'k' if sum(color[:3]) > 1.8 else 'w'
+            if hover_on == 'markers':
+                element.set_edgecolor(contrast_color)
+            annot.set_color(contrast_color)
             annot.set_text(element.get_label())
             annot.get_bbox_patch().set_facecolor(color)
             annot.get_bbox_patch().set_alpha(0.8)
@@ -118,9 +128,49 @@ def hover_annotation(event):
             fig.canvas.draw_idle()
             break
         element.set_linewidth(1)
+        if hover_on == 'markers':
+            element.set_edgecolor('w')
         if vis:
             annot.set_visible(False)
             fig.canvas.draw_idle()
+
+
+def _duration_units(duration):
+    """
+    Return the duration and the units.
+
+    :param duration: Duration in years
+    :type duration: float
+    :return: Duration, units
+    :rtype: float, str
+    """
+    units = 'years'
+    if duration < 1:
+        duration *= 12
+        units = 'months'
+    if duration < 1:
+        duration *= 30
+        units = 'days'
+    if duration < 1:
+        duration *= 24
+        units = 'hours'
+    if duration < 1:
+        duration *= 60
+        units = 'minutes'
+    if duration < 1:
+        duration *= 60
+        units = 'seconds'
+    return duration, units
+
+
+_short_units = {
+    'years': 'yrs',
+    'months': 'mos',
+    'days': 'days',
+    'hours': 'hrs',
+    'minutes': 'mins',
+    'seconds': 'secs',
+}
 
 
 def duration_string(family):
@@ -133,20 +183,137 @@ def duration_string(family):
     :rtype: str
     """
     duration = (family.endtime - family.starttime)/(365*24*60*60)
-    duration_str = f'{duration:.1f} yrs'
-    if duration < 1:
-        duration *= 12
-        duration_str = f'{duration:.1f} mos'
-    if duration < 1:
-        duration *= 30
-        duration_str = f'{duration:.1f} days'
-    if duration < 1:
-        duration *= 24
-        duration_str = f'{duration:.1f} hrs'
-    if duration < 1:
-        duration *= 60
-        duration_str = f'{duration:.1f} mins'
-    if duration < 1:
-        duration *= 60
-        duration_str = f'{duration:.1f} secs'
-    return duration_str
+    duration, units = _duration_units(duration)
+    return f'{duration:.1f} {_short_units[units]}'
+
+
+def family_colors(config, families):
+    """
+    Return the family colors, according to the colorby parameter.
+
+    :param config: Configuration object
+    :type config: Config
+    :param families: List of families
+    :type families: list
+    :return: List of colors, normalization object, colormap
+    :rtype: list, matplotlib.colors.Normalize, matplotlib.cm.ScalarMappable
+
+    :raises ValueError: If the colorby parameter is invalid or if lon0 and/or
+        lat0 are not specified when colorby is 'distance_from'
+    """
+    colorby = config.args.colorby
+    try:
+        cmap = cmaps[colorby]
+    except KeyError as e:
+        raise ValueError(f'Invalid value for "colorby": {colorby}') from e
+    # special cases
+    if colorby == 'family_number':
+        norm = colors.Normalize(vmin=-0.5, vmax=9.5)
+        fcolors = [cmap(norm(family.number % 10)) for family in families]
+        return fcolors, norm, cmap
+    if colorby == 'duration':
+        return _family_colors_duration(families, cmap)
+    if colorby == 'number_of_events':
+        return _family_color_number_of_events(families, cmap)
+    # general cases
+    if colorby == 'cumul_moment':
+        values = [family.cumul_moment for family in families]
+    elif colorby == 'cumul_slip':
+        values = [family.cumul_slip for family in families]
+    elif colorby == 'depth':
+        values = [family.depth for family in families]
+    elif colorby == 'distance_from':
+        lon0, lat0 = config.distance_from_lon, config.distance_from_lat
+        if lon0 is None or lat0 is None:
+            raise ValueError(
+                '"colorby" set to "distance_from", '
+                'but "distance_from_lon" and/or "distance_from_lat" '
+                'are not specified in the config file')
+        values = [family.distance_from(lon0, lat0) for family in families]
+        cmap.label = f'{cmap.label} ({lat0:.1f}°N,{lon0:.1f}°E) (km)'
+    elif colorby == 'latitude':
+        values = [family.lat for family in families]
+    elif colorby == 'longitude':
+        values = [family.lon for family in families]
+    elif colorby == 'slip_rate':
+        values = [family.slip_rate if family.slip_rate is not np.inf else np.nan for family in families]
+    elif colorby == 'time':
+        values = [family.starttime.matplotlib_date for family in families]
+    norm = colors.Normalize(vmin=min(values), vmax=max(values))
+    fcolors = [cmap(norm(value)) for value in values]
+    return fcolors, norm, cmap
+
+
+def _family_color_number_of_events(families, cmap):
+    """
+    Return the family colors, according to the number of events.
+
+    :param families: List of families
+    :type families: list
+    :param cmap: Colormap
+    :type cmap: matplotlib.colors.Colormap
+    :return: List of colors, normalization object, colormap
+    :rtype: list, matplotlib.colors.Normalize, matplotlib.cm.ScalarMappable
+    """
+    values = [len(family) for family in families]
+    boundaries = np.arange(min(values), max(values)+1)
+    if len(boundaries) % 2 == 0:
+        boundaries = np.append(boundaries, max(values)+1)
+    boundaries = boundaries - 0.5
+    norm = colors.BoundaryNorm(boundaries, cmap.N)
+    return [cmap(norm(value)) for value in values], norm, cmap
+
+
+def _family_colors_duration(families, cmap):
+    """
+    Return the family colors, according to the duration.
+
+    :param families: List of families
+    :type families: list
+    :param cmap: Colormap
+    :type cmap: matplotlib.colors.Colormap
+    :return: List of colors, normalization object, colormap
+    :rtype: list, matplotlib.colors.Normalize, matplotlib.cm.ScalarMappable
+    """
+    durations = [family.duration for family in families]
+    max_duration = max(durations)
+    max_duration_new_units, units = _duration_units(max_duration)
+    multiplier = max_duration_new_units/max_duration
+    min_duration_new_units = min(durations)*multiplier
+    norm = colors.Normalize(
+        vmin=min_duration_new_units, vmax=max_duration_new_units)
+    fcolors = [cmap(norm(duration*multiplier)) for duration in durations]
+    cmap.label = f'{cmap.label} ({units})'
+    return fcolors, norm, cmap
+
+
+def plot_colorbar(config, fig, ax, cmap, norm):
+    """
+    Add a colorbar to a plot.
+
+    :param config: Configuration object
+    :type config: Config
+    :param fig: Matplotlib figure
+    :type fig: matplotlib.figure.Figure
+    :param ax: Matplotlib axis
+    :type ax: matplotlib.axes.Axes
+    :param cmap: Colormap
+    :type cmap: matplotlib.colors.Colormap
+    :param norm: Normalization object
+    :type norm: matplotlib.colors.Normalize
+    """
+    colorby = config.args.colorby
+    sm = cm.ScalarMappable(cmap=cmap, norm=norm)
+    cbar_ticks = None
+    if colorby == 'family_number':
+        cbar_ticks = range(10)
+    elif colorby == 'number_of_events':
+        cbar_ticks = norm.boundaries + 0.5
+    cbar = fig.colorbar(sm, ticks=cbar_ticks, pad=0.1, ax=ax)
+    # turn off minor ticks
+    cbar.ax.yaxis.set_tick_params(which='minor', size=0)
+    cbar.ax.set_zorder(-1)
+    if colorby == 'time':
+        format_time_axis(cbar.ax, which='yaxis', grid=False)
+    with contextlib.suppress(AttributeError):
+        cbar.ax.set_ylabel(cmap.label)